requirements.txt
google-generativeai
openai>=1.20.0
pandas==2.2.2
protobuf<6,>=3.20.2
pydantic<2.0.0
Pympler==1.0.1
click>=8.1.7
ollama
fastapi
torch
transformers
accelerate
anthropic
python-dotenv

langchain-openai==0.1.7
langchain-ollama
chromadb==0.4.24
jsonlines
langchain<0.3.0
langchain-community<0.3.0
